:recycle: Rename emptyStateV2 to emptyState
