+++ src/components/emptyState/emptyState.css
@@ -1,46 +1,29 @@
-@import '../app/variables.css';
+@import '../app/variablesV2.css';
+@import '../app/mixins.css';
 
-:root {
-  --float-animation-distance: 10px;
-}
+.wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  box-sizing: border-box;
+  padding: 20px;
 
-@keyframes float {
-  0% {
-    margin-top: 0px;
-    margin-bottom: var(--float-animation-distance);
+  & > h1 {
+    @mixin headingNormal;
+
+    color: var(--color-maastricht-blue);
+    margin: 20px 0 0;
   }
 
-  100% {
-    margin-top: var(--float-animation-distance);
-    margin-bottom: 0px;
-  }
-}
+  & > p {
+    @mixin contentLarge;
 
-.emptyState {
-  max-width: 255px;
-  margin: auto;
-  text-align: center;
-
-  & img {
-    animation: float 1500ms 0ms infinite alternate ease-in-out;
-  }
-
-  & h2 {
-    margin: 10px;
-  }
-
-  & p {
-    line-height: 26px;
-    color: var(--color-grayscale-dark);
-  }
-}
-
-@media (--small-viewport) {
-  .emptyState {
-    margin: 50px auto;
-
-    & p {
-      font-size: 14px;
-    }
+    max-width: 267px;
+    color: var(--color-slate-gray);
+    text-align: center;
+    margin: 20px 0;
   }
 }
