--- src/components/emptyStateV2/emptyState.css
@@ -1,29 +1,0 @@
-@import '../app/variablesV2.css';
-@import '../app/mixins.css';
-
-.wrapper {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  box-sizing: border-box;
-  padding: 20px;
-
-  & > h1 {
-    @mixin headingNormal;
-
-    color: var(--color-maastricht-blue);
-    margin: 20px 0 0;
-  }
-
-  & > p {
-    @mixin contentLarge;
-
-    max-width: 267px;
-    color: var(--color-slate-gray);
-    text-align: center;
-    margin: 20px 0;
-  }
-}
