--- src/components/emptyStateV2/index.js
@@ -1,11 +1,0 @@
-// istanbul ignore file
-import React from 'react';
-import styles from './emptyState.css';
-
-const EmptyState = ({ className = '', children }) => (
-  <div className={`${styles.wrapper} ${className} empty-state`}>
-    {children}
-  </div>
-);
-
-export default EmptyState;
