+++ src/components/newsFeed/newsFeed.js
@@ -2,7 +2,7 @@ import React from 'react';
 import styles from './newsFeed.css';
 import News from './news';
 import Box from '../box';
-import EmptyState from '../emptyStateV2';
+import EmptyState from '../emptyState';
 import Icon from '../toolbox/icon';
 
 class NewsFeed extends React.Component {
